♻️ refactor: report 상수, 타입 분리
